requirements.txt
openai-whisper==20231117
colorama==0.4.6